Extract theme import parser and export logic into ThemeSerializationService (Plan 00067)
